Reorg editor components
