+++ services/user/Social/ui/src/components/index.ts
@@ -1,5 +1,6 @@
 export * from "./account-switcher";
 export * from "./mail-display";
 export * from "./mail-list";
+export * from "./markdown-editor";
 export * from "./nav";
 export * from "./nav-menu";
