+++ services/user/Social/ui/src/routes/editor.tsx
@@ -1,18 +1,6 @@
-import {
-    defaultValueCtx,
-    editorViewOptionsCtx,
-    Editor as MilkdownEditor,
-    rootCtx,
-} from "@milkdown/core";
-import { nord } from "@milkdown/theme-nord";
-import { Milkdown, MilkdownProvider, useEditor } from "@milkdown/react";
-import { commonmark } from "@milkdown/preset-commonmark";
-import { gfm } from "@milkdown/preset-gfm";
-import { listItemBlockComponent } from "@milkdown/components/list-item-block";
-import { listener, listenerCtx } from "@milkdown/plugin-listener";
+import { MilkdownProvider } from "@milkdown/react";
 
-import "@milkdown/theme-nord/style.css";
-import "../styles/editor.css";
+import { MarkdownEditor } from "@components";
 import { useLocalStorage } from "@hooks";
 
 const defaultMarkdown = `# Milkdown
@@ -130,44 +118,6 @@ Have fun!
 
 `;
 
-export const MilkdownEditorComponent = ({
-    initialValue,
-    updateMarkdown,
-    readOnly = false,
-}: {
-    initialValue: string;
-    updateMarkdown: (value: string) => void;
-    readOnly?: boolean;
-}) => {
-    const { get } = useEditor((root) =>
-        MilkdownEditor.make()
-            .config(nord)
-            .config((ctx) => {
-                ctx.set(rootCtx, root);
-                ctx.set(defaultValueCtx, initialValue);
-            })
-            .config((ctx) => {
-                const listener = ctx.get(listenerCtx);
-                listener.markdownUpdated((_ctx, md, prevMd) => {
-                    if (md === prevMd) return;
-                    updateMarkdown(md);
-                });
-            })
-            .config((ctx) => {
-                ctx.update(editorViewOptionsCtx, (prev) => ({
-                    ...prev,
-                    editable: () => !readOnly,
-                }));
-            })
-            .use(listener)
-            .use(commonmark)
-            .use(gfm)
-            .use(listItemBlockComponent),
-    );
-
-    return <Milkdown />;
-};
-
 export function Editor() {
     const [markdown, setMarkdown] = useLocalStorage(
         "markdown",
@@ -177,7 +127,7 @@ export function Editor() {
     return (
         <MilkdownProvider>
             {markdown ? (
-                <MilkdownEditorComponent
+                <MarkdownEditor
                     initialValue={markdown}
                     updateMarkdown={setMarkdown}
                 />
